backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.14.0,<3.15
django-cors-headers>=4.3.0,<4.4
Pillow>=10.0.0,<11.0  # Pour ImageField dans vos modèles